requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
google-genai
openai
gtts
python-multipart
